LayoutTests/inspector-obsolete/http-tests/inspect-iframe-from-different-domain-expected.txt

CONSOLE MESSAGE: line 6: IFRAME HOST: localhost:8000

Tests that style properties of elements in iframes loaded from domain different from the main document domain can be inspected. See bug 31587.

background: green;
